backend/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
PyJWT==2.8.0
cryptography==41.0.5
python-jose==3.3.0
authlib==1.2.1
requests==2.31.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
gunicorn==21.2.0
pytest==7.4.3
